Consolidate Account and UserAgent tests
